+++ sentry-kotlin-multiplatform/src/androidMain/kotlin/io/sentry/kotlin/multiplatform/Context.android.kt
@@ -6,18 +6,11 @@ import android.content.Context
 import android.database.Cursor
 import android.net.Uri
 import io.sentry.android.core.SentryAndroid
-import io.sentry.kotlin.multiplatform.extensions.toAndroidSentryOptionsCallback
-
-internal actual fun initSentry(configuration: OptionsConfiguration) {
-    val options = SentryOptions()
-    configuration.invoke(options)
-    SentryAndroid.init(applicationContext, options.toAndroidSentryOptionsCallback())
-}
-
-internal lateinit var applicationContext: Context
-    private set
 
 public actual typealias Context = Context
+
+internal var applicationContext: Context? = null
+    private set
 
 /**
  * A ContentProvider that does NOT store or provide any data for read or write operations.