+++ sentry-kotlin-multiplatform/src/androidUnitTest/kotlin/io/sentry/kotlin/multiplatform/PlatformOptions.android.kt
@@ -2,6 +2,7 @@ package io.sentry.kotlin.multiplatform
 
 import io.sentry.android.core.SentryAndroidOptions
 import io.sentry.kotlin.multiplatform.extensions.toAndroidSentryOptionsCallback
+import io.sentry.kotlin.multiplatform.utils.fakeDsn
 import kotlin.test.assertEquals
 
 actual interface PlatformOptions : CommonPlatformOptions {
@@ -74,4 +75,8 @@ actual fun PlatformOptions.assertPlatformSpecificOptions(options: SentryOptions)
     assertEquals(attachViewHierarchy, options.attachViewHierarchy)
     assertEquals(isAnrEnabled, options.isAnrEnabled)
     assertEquals(anrTimeoutIntervalMillis, options.anrTimeoutIntervalMillis)
+}
+
+actual fun createSentryPlatformOptionsConfiguration(): PlatformOptionsConfiguration = {
+    it.dsn = fakeDsn
 }