+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/SentryBridge.apple.kt
@@ -1,10 +1,13 @@
 package io.sentry.kotlin.multiplatform
 
+import NSException.Sentry.SentryEvent
+import PrivateSentrySDKOnly.Sentry.PrivateSentrySDKOnly
 import cocoapods.Sentry.SentrySDK
 import io.sentry.kotlin.multiplatform.extensions.toCocoaBreadcrumb
 import io.sentry.kotlin.multiplatform.extensions.toCocoaUser
 import io.sentry.kotlin.multiplatform.extensions.toCocoaUserFeedback
 import io.sentry.kotlin.multiplatform.nsexception.asNSException
+import io.sentry.kotlin.multiplatform.nsexception.dropKotlinCrashEvent
 import io.sentry.kotlin.multiplatform.protocol.Breadcrumb
 import io.sentry.kotlin.multiplatform.protocol.SentryId
 import io.sentry.kotlin.multiplatform.protocol.User
@@ -14,16 +17,53 @@ import platform.Foundation.NSException
 
 public actual abstract class Context
 
-internal expect fun initSentry(configuration: OptionsConfiguration)
+// Since the function is the same on all apple platforms, we don't split it into expect/actual
+// like on JVM and Android, we may do that later on if needed.
+internal actual fun SentryPlatformOptions.prepareForInit() {
+    val cocoa = this as? CocoaSentryOptions
+    val existingBeforeSend = cocoa?.beforeSend
+    val modifiedBeforeSend: (CocoaSentryEvent?) -> CocoaSentryEvent? = beforeSend@{ event ->
+        // Return early if the user's beforeSend returns null
+        if (existingBeforeSend?.invoke(event) == null) {
+            return@beforeSend null
+        }
 
-internal actual object SentryBridge {
+        val cocoaName = BuildKonfig.SENTRY_COCOA_PACKAGE_NAME
+        val cocoaVersion = BuildKonfig.SENTRY_COCOA_VERSION
 
+        val sdk = event?.sdk?.toMutableMap() ?: mutableMapOf()
+        val packages = sdk["packages"] as? MutableList<Map<String, String>> ?: mutableListOf()
+
+        packages.add(mapOf("name" to cocoaName, "version" to cocoaVersion))
+        sdk["packages"] = packages
+        event?.sdk = sdk
+
+        dropKotlinCrashEvent(event as SentryEvent?) as CocoaSentryEvent?
+    }
+
+    cocoa?.setBeforeSend(modifiedBeforeSend)
+
+    PrivateSentrySDKOnly.setSdkName(BuildKonfig.SENTRY_KMP_COCOA_SDK_NAME, BuildKonfig.VERSION_NAME)
+}
+
+internal actual class SentryBridge actual constructor(private val sentryInstance: SentryInstance) {
     actual fun init(context: Context, configuration: OptionsConfiguration) {
-        initSentry(configuration)
+        init(configuration)
     }
 
     actual fun init(configuration: OptionsConfiguration) {
-        initSentry(configuration)
+        val options = SentryOptions()
+        configuration.invoke(options)
+        initWithPlatformOptions(options.toPlatformOptionsConfiguration())
+    }
+
+    actual fun initWithPlatformOptions(configuration: PlatformOptionsConfiguration) {
+        val finalConfiguration: PlatformOptionsConfiguration = {
+            configuration(it)
+            // We modify beforeSend so we need this to run after the user's configuration
+            it.prepareForInit()
+        }
+        sentryInstance.init(finalConfiguration)
     }
 
     actual fun captureMessage(message: String): SentryId {