+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SentryOptionsExtensions.jvm.kt
@@ -1,6 +1,5 @@
 package io.sentry.kotlin.multiplatform.extensions
 
-import io.sentry.kotlin.multiplatform.BuildKonfig
 import io.sentry.kotlin.multiplatform.JvmSentryOptions
 import io.sentry.kotlin.multiplatform.SentryEvent
 import io.sentry.kotlin.multiplatform.SentryOptions
@@ -8,19 +7,8 @@ import io.sentry.kotlin.multiplatform.SentryOptions
 internal fun SentryOptions.toJvmSentryOptionsCallback(): (JvmSentryOptions) -> Unit = {
     it.applyJvmBaseOptions(this)
 
-    // Apply JVM specific options
-    it.sdkVersion?.name = sdk?.name ?: BuildKonfig.SENTRY_KMP_JAVA_SDK_NAME
-    it.sdkVersion?.version = sdk?.version ?: BuildKonfig.VERSION_NAME
-
     sdk?.packages?.forEach { sdkPackage ->
         it.sdkVersion?.addPackage(sdkPackage.name, sdkPackage.version)
-    }
-
-    if (it.sdkVersion?.packages?.none { it.name == BuildKonfig.SENTRY_JAVA_PACKAGE_NAME } == true) {
-        it.sdkVersion?.addPackage(
-            BuildKonfig.SENTRY_JAVA_PACKAGE_NAME,
-            BuildKonfig.SENTRY_JAVA_VERSION
-        )
     }
 }