+++ sentry-kotlin-multiplatform/src/iosMain/kotlin/io/sentry/kotlin/multiplatform/SentryPlatformOptions.ios.kt
@@ -1,0 +1,8 @@
+package io.sentry.kotlin.multiplatform
+
+import io.sentry.kotlin.multiplatform.extensions.toIosOptionsConfiguration
+
+public actual typealias SentryPlatformOptions = cocoapods.Sentry.SentryOptions
+
+internal actual fun SentryOptions.toPlatformOptionsConfiguration(): PlatformOptionsConfiguration =
+    toIosOptionsConfiguration()