+++ sentry-kotlin-multiplatform/src/jvmMain/kotlin/io/sentry/kotlin/multiplatform/Context.jvm.kt
@@ -1,0 +1,4 @@
+package io.sentry.kotlin.multiplatform
+
+// The context is unused here and only implemented to satisfy the expect
+public actual abstract class Context