+++ sentry-kotlin-multiplatform/src/jvmTest/kotlin/io/sentry/kotlin/multiplatform/PlatformOptions.jvm.kt
@@ -1,6 +1,7 @@
 package io.sentry.kotlin.multiplatform
 
 import io.sentry.kotlin.multiplatform.extensions.toJvmSentryOptionsCallback
+import io.sentry.kotlin.multiplatform.utils.fakeDsn
 
 actual interface PlatformOptions : CommonPlatformOptions
 
@@ -50,4 +51,8 @@ actual fun createPlatformOptions(): PlatformOptions = SentryJvmOptionsWrapper(Jv
 
 actual fun PlatformOptions.assertPlatformSpecificOptions(options: SentryOptions) {
     // no-op
+}
+
+actual fun createSentryPlatformOptionsConfiguration(): PlatformOptionsConfiguration = {
+    it.dsn = fakeDsn
 }